+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.57](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.56...@codecademy/gamut-kit@0.6.57) (2022-02-04)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.56](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.55...@codecademy/gamut-kit@0.6.56) (2022-02-04)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.6.55](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.54...@codecademy/gamut-kit@0.6.55) (2022-02-03)
 
 **Note:** Version bump only for package @codecademy/gamut-kit